requirements.txt
absl-py==0.13.0
amqp==2.6.1
astor==0.8.1
backcall==0.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==4.4.7
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cycler==0.10.0
decorator==4.4.2
Flask==2.0.1
flower==0.9.7
gast==0.4.0
gevent==21.1.2
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.38.1
h5py==2.10.0
humanize==3.9.0
imageio==2.9.0
importlib-metadata==4.5.0
ipython==7.25.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kombu==4.6.11
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mrcnn==0.2
networkx==2.5.1
numpy==1.21.0
numpyencoder==0.3.0
opencv-python==4.5.2.54
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
redis==3.5.3
scikit-image==0.17.2
scipy==1.5.2
six==1.16.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tifffile==2021.6.14
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
vine==1.3.0
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
